requirements.txt
git+https://github.com/Rapptz/discord.py
requests
beautifulsoup4
python-dotenv~=0.15.0
requests~=2.25.0
